pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-framework"
version = "0.1.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
description = "A modular agent framework"
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.6.1,<3.0.0",
    "python-dateutil>=2.8.2,<3.0.0",
    "typing-extensions>=4.9.0,<5.0.0",
    "uuid>=1.30,<2.0.0",
    "openai>=1.12.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[tool.setuptools.packages.find]
include = ["agent_framework*"] 